requirements.txt
requests==2.26.0
xmltojson==0.2.0
xmltodict==0.12.0
utils==1.0.1
prettytable==2.2.0
tabulate==0.8.9
python-decouple==3.4